+++ src/lib/accordions/AccordionFlush.svelte
@@ -8,8 +8,9 @@
 	export let id: string = '';
 	export let btnClass: string =
 		'flex justify-between items-center py-5 w-full font-medium text-left text-gray-500 border-b border-gray-200 dark:border-gray-700 dark:text-gray-400';
-
 	export let slotClass: string = 'py-5 border-b border-gray-200 dark:border-gray-700';
+	export let iconSize: number = 24;
+	export let iconClass: string = 'text-gray-500 sm:w-6 sm:h-6 dark:text-gray-300';
 	export let isOpen: boolean = false;
 	export let icons: AccordionIconType = {
 		up: ChevronUp,
@@ -25,9 +26,6 @@
 	const handleToggle = (id: string) => {
 		isOpen = !isOpen;
 	};
-
-	export let iconSize: number = 24;
-	export let iconClass: string = 'text-gray-500 sm:w-6 sm:h-6 dark:text-gray-300';
 </script>
 
 <h2 aria-expanded={isOpen}>