+++ src/lib/utils/clickOutside.ts
@@ -1,4 +1,4 @@
-export function clickOutside(node: HTMLElement, callback: () => void) {
+export const clickOutside=(node: HTMLElement, callback: () => void)=> {
   const handleClick = (event: MouseEvent) => {
     if (!event?.target) return;
     if (node && !node.contains(event.target as Node) && !event.defaultPrevented) {